+++ example/lib/page/player/video_play_page.dart
@@ -24,13 +24,13 @@ class _VideoPlayPageState extends State<VideoPlayPage> {
 
   void initController() {
     List<String> urls = [
-      'https://test-videos.co.uk/vids/jellyfish/mkv/1080/Jellyfish_1080_10s_30MB.mkv',
-      'https://www.sample-videos.com/video321/3gp/240/big_buck_bunny_240p_30mb.3gp',
-      'http://vjs.zencdn.net/v/oceans.mp4',
-      'https://user-images.githubusercontent.com/28951144/229373695-22f88f13-d18f-4288-9bf1-c3e078d83722.mp4',
+      'https://vv.jisuzyv.com/play/DbDGZ8ka/index.m3u8',
       'https://test-streams.mux.dev/x36xhzz/url_6/193039199_mp4_h264_aac_hq_7.m3u8',
       'https://customer-fzuyyy7va6ohx90h.cloudflarestream.com/861279ab37d84dbfbf3247322fbcfc63/manifest/video.m3u8',
-      'https://vv.jisuzyv.com/play/DbDGZ8ka/index.m3u8',
+      'http://vjs.zencdn.net/v/oceans.mp4',
+      'https://user-images.githubusercontent.com/28951144/229373695-22f88f13-d18f-4288-9bf1-c3e078d83722.mp4',
+      'https://www.sample-videos.com/video321/3gp/240/big_buck_bunny_240p_30mb.3gp',
+      'https://test-videos.co.uk/vids/jellyfish/mkv/1080/Jellyfish_1080_10s_30MB.mkv',
     ];
     var url = urls[0];
     String localUri = url.toLocalUrl();
@@ -45,6 +45,7 @@ class _VideoPlayPageState extends State<VideoPlayPage> {
         setState(() {});
         timer = Timer(const Duration(milliseconds: 200), () {
           _controller.play();
+          setState(() {});
         });
       });
   }
@@ -59,15 +60,41 @@ class _VideoPlayPageState extends State<VideoPlayPage> {
           ? Stack(
               alignment: Alignment.bottomCenter,
               children: [
-                AspectRatio(
-                  aspectRatio: _controller.value.aspectRatio,
-                  child: VideoPlayer(_controller),
+                GestureDetector(
+                  onTap: () {
+                    _controller.value.isPlaying
+                        ? _controller.pause()
+                        : _controller.play();
+                    setState(() {});
+                  },
+                  child: AspectRatio(
+                    aspectRatio: _controller.value.aspectRatio,
+                    child: VideoPlayer(_controller),
+                  ),
                 ),
                 SizedBox(
                   height: 20,
                   child: VideoProgressIndicator(
                     _controller,
                     allowScrubbing: true,
+                  ),
+                ),
+                Positioned.fill(
+                  child: Offstage(
+                    offstage: _controller.value.isPlaying,
+                    child: GestureDetector(
+                      onTap: () {
+                        _controller.value.isPlaying
+                            ? _controller.pause()
+                            : _controller.play();
+                        setState(() {});
+                      },
+                      child: Icon(
+                        Icons.play_circle,
+                        color: Colors.white,
+                        size: 40,
+                      ),
+                    ),
                   ),
                 ),
               ],
@@ -87,15 +114,10 @@ class _VideoPlayPageState extends State<VideoPlayPage> {
             ),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
-          setState(() {
-            _controller.value.isPlaying
-                ? _controller.pause()
-                : _controller.play();
-          });
+          List<DownloadTask> tasks = VideoProxy.downloadManager.allTasks;
+          logW("task: ${tasks.map((e) => e.toString()).toString()}");
         },
-        child: Icon(
-          _controller.value.isPlaying ? Icons.pause : Icons.play_arrow,
-        ),
+        child: Icon(Icons.code),
       ),
     );
   }